+++ src/cxl_manager/cxl_manager.h
@@ -56,7 +56,7 @@ struct alignas(64) TInode{
 };
 
 struct NonCriticalMessageHeader{
-	int client_id;
+	uint16_t client_id;
 	size_t client_order;
 	size_t size;
 	size_t paddedSize;
@@ -76,7 +76,7 @@ struct alignas(64) MessageHeader{
 	size_t logical_offset;
 	unsigned long long int next_msg_diff; // Relative to message_header, not cxl_addr_
 	volatile size_t total_order;
-	uint32_t client_id;
+	uint16_t client_id;
 	uint32_t client_order;
 	size_t size;
 	volatile size_t paddedSize; // This include message+padding+header size
